requirements.txt
streamlit
torch
torchvision
opencv-python-headless
albumentations
pillow
seaborn
numpy
matplotlib
protobuf
